requirements.txt
Django==1.3.1
django-tastypie==0.9.11
django-social-auth==0.6.3
psycopg2==2.4.4
requests==0.10.6
South==0.7.3
